requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
openai==0.26.0
pymongo==4.3.3
psycopg2-binary==2.9.6
pinecone-client==2.1.0
requests==2.28.1
python-dotenv==0.20.0
